+++ src/errors/network/ForbiddenError.ts
