+++ src/errors/network/NetworkError.ts
