+++ src/errors/network/NotFoundError.ts
