+++ src/errors/network/ValidationError.ts
